+++ packages/ast-spec/src/declaration/ImportDeclaration/fixtures/defer/fixture.ts
@@ -1,0 +1,1 @@
+import defer * as mod from 'mod';