+++ packages/typescript-estree/src/check-syntax-errors.ts
@@ -3,6 +3,7 @@ import * as ts from 'typescript';
 import type { TSNode } from './ts-estree';
 
 import { checkModifiers } from './check-modifiers';
+import { getImportClausePhaseModifier } from './getImportClausePhaseModifier';
 import {
   isValidAssignmentTarget,
   createError,
@@ -247,6 +248,29 @@ export function checkSyntaxError(
         );
       }
 
+      const isDefinite = !!node.exclamationToken;
+
+      if (isDefinite && isAbstract) {
+        throw createError(
+          node.exclamationToken,
+          `A definite assignment assertion '!' is not permitted in this context.`,
+        );
+      }
+
+      if (isDefinite && !node.type) {
+        throw createError(
+          node,
+          `Declarations with definite assignment assertions must also have type annotations.`,
+        );
+      }
+
+      if (isDefinite && node.initializer) {
+        throw createError(
+          node,
+          `Declarations with initializers cannot also have definite assignment assertions.`,
+        );
+      }
+
       if (
         node.name.kind === SyntaxKind.StringLiteral &&
         node.name.text === 'constructor'
@@ -366,11 +390,8 @@ export function checkSyntaxError(
     case SyntaxKind.ImportDeclaration: {
       const { importClause } = node;
       if (
-        // TODO swap to `phaseModifier` once we add support for `import defer`
-        // https://github.com/estree/estree/issues/328
-        // eslint-disable-next-line @typescript-eslint/no-deprecated
-        importClause?.isTypeOnly &&
-        importClause.name &&
+        getImportClausePhaseModifier(importClause) === 'type' &&
+        importClause?.name &&
         importClause.namedBindings
       ) {
         throw createError(